Get dependencies of a package in Bioconductor

This function queries Bioconductor for dependencies of the selected package for given Bioconductor release.

Usage

wood_bioc_dependencies(package, release = "release")

Arguments

package

character(1)
Name of a package to query for.

release

character(1)
Bioconductor release to query for. Can be either a numeric code (like 3.12) or one of two keywords: "release" and "devel".

Value

A data frame with three columns, all in string format:

package (package name),

version (minimum version requirement or NA if none),

type (dependency type, e.g. "Imports").

See also

Examples

# \donttest{
wood_bioc_dependencies("Biostrings")
#> <dependencies>
#>   Depends:     R (>= 4.0.0)
#>   Depends:     methods
#>   Depends:     BiocGenerics (>= 0.37.0)
#>   Depends:     S4Vectors (>= 0.27.12)
#>   Depends:     IRanges (>= 2.31.2)
#>   Depends:     XVector (>= 0.37.1)
#>   Depends:     GenomeInfoDb
#>   Imports:     methods
#>   Imports:     utils
#>   Imports:     grDevices
#>   Imports:     graphics
#>   Imports:     stats
#>   Imports:     crayon
#>   LinkingTo:   S4Vectors
#>   LinkingTo:   IRanges
#>   LinkingTo:   XVector
#>   Suggests:    BSgenome (>= 1.13.14)
#>   Suggests:    BSgenome.Celegans.UCSC.ce2 (>= 1.3.11)
#>   Suggests:    BSgenome.Dmelanogaster.UCSC.dm3 (>= 1.3.11)
#>   Suggests:    BSgenome.Hsapiens.UCSC.hg18
#>   Suggests:    drosophila2probe
#>   Suggests:    hgu95av2probe
#>   Suggests:    hgu133aprobe
#>   Suggests:    GenomicFeatures (>= 1.3.14)
#>   Suggests:    hgu95av2cdf
#>   Suggests:    affy (>= 1.41.3)
#>   Suggests:    affydata (>= 1.11.5)
#>   Suggests:    RUnit
#>   Enhances:    Rmpi

# Will dependencies change?
wood_bioc_dependencies("Biostrings", "devel")
#> <dependencies>
#>   Depends:     R (>= 4.0.0)
#>   Depends:     methods
#>   Depends:     BiocGenerics (>= 0.37.0)
#>   Depends:     S4Vectors (>= 0.27.12)
#>   Depends:     IRanges (>= 2.31.2)
#>   Depends:     XVector (>= 0.37.1)
#>   Depends:     GenomeInfoDb
#>   Imports:     methods
#>   Imports:     utils
#>   Imports:     grDevices
#>   Imports:     graphics
#>   Imports:     stats
#>   Imports:     crayon
#>   LinkingTo:   S4Vectors
#>   LinkingTo:   IRanges
#>   LinkingTo:   XVector
#>   Suggests:    BSgenome (>= 1.13.14)
#>   Suggests:    BSgenome.Celegans.UCSC.ce2 (>= 1.3.11)
#>   Suggests:    BSgenome.Dmelanogaster.UCSC.dm3 (>= 1.3.11)
#>   Suggests:    BSgenome.Hsapiens.UCSC.hg18
#>   Suggests:    drosophila2probe
#>   Suggests:    hgu95av2probe
#>   Suggests:    hgu133aprobe
#>   Suggests:    GenomicFeatures (>= 1.3.14)
#>   Suggests:    hgu95av2cdf
#>   Suggests:    affy (>= 1.41.3)
#>   Suggests:    affydata (>= 1.11.5)
#>   Suggests:    RUnit
#>   Enhances:    Rmpi
# And what about dependencies in the past?
wood_bioc_dependencies("Biostrings", "2.10")
#> <dependencies>
#>   Depends:     R (>= 2.8.0)
#>   Depends:     methods
#>   Depends:     BiocGenerics (>= 0.1.2)
#>   Depends:     IRanges (>= 1.13.6)
#>   Imports:     graphics
#>   Imports:     methods
#>   Imports:     stats
#>   Imports:     utils
#>   Imports:     BiocGenerics
#>   Imports:     IRanges
#>   LinkingTo:   IRanges
#>   Suggests:    BSgenome (>= 1.13.14)
#>   Suggests:    BSgenome.Celegans.UCSC.ce2 (>= 1.3.11)
#>   Suggests:    BSgenome.Dmelanogaster.UCSC.dm3 (>= 1.3.11)
#>   Suggests:    drosophila2probe
#>   Suggests:    hgu95av2probe
#>   Suggests:    hgu133aprobe
#>   Suggests:    GenomicFeatures (>= 1.3.14)
#>   Suggests:    hgu95av2cdf
#>   Suggests:    affy
#>   Suggests:    affydata (>= 1.11.5)
#>   Suggests:    RUnit
#>   Enhances:    Rmpi
wood_bioc_dependencies("Biostrings", "1.5")
#> <dependencies>
#>   Depends:     R (>= 1.8.0)
# }