List available packages in Bioconductor

This function queries Bioconductor for a list of available packages. Current and previous releases can be queried, up to Bioconductor 1.5.

Usage

wood_bioc_packages(release = "release")

Arguments

release

character(1)
Bioconductor release to query for. Can be either a numeric code (like 3.12) or one of two keywords: "release" and "devel".

Value

A character vector of available packages.

See also

Examples

# \donttest{
head(wood_bioc_packages())
#> [1] "a4"          "a4Base"      "a4Classif"   "a4Core"      "a4Preproc"  
#> [6] "a4Reporting"
# is the same as
head(wood_bioc_packages("release"))
#> [1] "a4"          "a4Base"      "a4Classif"   "a4Core"      "a4Preproc"  
#> [6] "a4Reporting"

# A glimpse in the future:
head(wood_bioc_packages("devel"))
#> [1] "a4"          "a4Base"      "a4Classif"   "a4Core"      "a4Preproc"  
#> [6] "a4Reporting"
# Previous versions can be queried as well:
head(wood_bioc_packages("3.9"))
#> [1] "a4"          "a4Base"      "a4Classif"   "a4Core"      "a4Preproc"  
#> [6] "a4Reporting"
# The oldest available version is 1.5:
wood_bioc_packages("1.5")
#>  [1] "aCGH"                 "affy"                 "affycomp"            
#>  [4] "affydata"             "affylmGUI"            "affypdnn"            
#>  [7] "affyPLM"              "altcdfenvs"           "annaffy"             
#> [10] "AnnBuilder"           "annotate"             "apComplex"           
#> [13] "arrayMagic"           "arrayQuality"         "bim"                 
#> [16] "Biobase"              "Biostrings"           "ChromoViz"           
#> [19] "convert"              "ctc"                  "daMA"                
#> [22] "DEDS"                 "DNAcopy"              "DynDoc"              
#> [25] "EBarrays"             "ecolitk"              "edd"                 
#> [28] "exprExternal"         "externalVector"       "factDesign"          
#> [31] "gcrma"                "genArise"             "genefilter"          
#> [34] "geneplotter"          "GeneSpring"           "GeneTraffic"         
#> [37] "GeneTS"               "GLAD"                 "globaltest"          
#> [40] "goCluster"            "GOstats"              "goTools"             
#> [43] "gpls"                 "graph"                "GraphAT"             
#> [46] "gtkWidgets"           "HEM"                  "hexbin"              
#> [49] "hopach"               "Icens"                "impute"              
#> [52] "iSPlot"               "KEGGSOAP"             "limma"               
#> [55] "limmaGUI"             "LPE"                  "makecdfenv"          
#> [58] "marray"               "matchprobes"          "MeasurementError.cor"
#> [61] "MergeMaid"            "MLInterfaces"         "msbase"              
#> [64] "multtest"             "nnNorm"               "OLIN"                
#> [67] "OLINgui"              "pairseqsim"           "pamr"                
#> [70] "pickgene"             "prada"                "PROcess"             
#> [73] "qvalue"               "rama"                 "RBGL"                
#> [76] "Rdbi"                 "RdbiPgSQL"            "reposTools"          
#> [79] "Resourcerer"          "Rgraphviz"            "RMAGEML"             
#> [82] "ROC"                  "RSNPper"              "Ruuid"               
#> [85] "SAGElyzer"            "siggenes"             "simpleaffy"          
#> [88] "SNAData"              "splicegear"           "stam"                
#> [91] "stepNorm"             "tkWidgets"            "twilight"            
#> [94] "vsn"                  "webbioc"              "widgetInvoke"        
#> [97] "widgetTools"         
# }