Get current package version on GitLab

This function queries GitLab for the code of the current package version. This may reference a non-tagged commit; for the analysis of tagged commits only, see wood_gitlab_versions().

Usage

wood_gitlab_latest(package, user)

Arguments

package

character(1)
Name of a package to query for.

user

character(1)
Name of a user or organization.

Value

A character vector of version codes.

See also

Examples

if (interactive()) {
  # Latest version code is returned
  wood_gitlab_latest("rock", "r-packages")

  # To get the latest *tagged* version code instead, use:
  wood_gitlab_versions("rock", "r-packages") |>
    versionsort::ver_latest()
}