List locally available packages

This function searches for installed packages inside specified library paths.

Usage

wood_local_packages(paths = .libPaths()[1])

Arguments

paths

character()
Paths to local libraries, by default the first element of .libPaths(). If equal to "all", uses all elements of .libPaths().

Value

A character vector of available packages.

See also

Examples

wood_local_packages()
#>  [1] "R6"          "_cache"      "askpass"     "base64enc"   "brio"       
#>  [6] "bslib"       "cachem"      "callr"       "cli"         "clipr"      
#> [11] "covr"        "crayon"      "credentials" "curl"        "desc"       
#> [16] "diffobj"     "digest"      "downlit"     "evaluate"    "fansi"      
#> [21] "fastmap"     "fontawesome" "fs"          "gert"        "gh"         
#> [26] "gitcreds"    "glue"        "highr"       "htmltools"   "httptest2"  
#> [31] "httr"        "httr2"       "ini"         "jquerylib"   "jsonlite"   
#> [36] "knitr"       "lazyeval"    "lifecycle"   "magrittr"    "memoise"    
#> [41] "mime"        "openssl"     "pillar"      "pkgbuild"    "pkgconfig"  
#> [46] "pkgdown"     "pkgload"     "praise"      "processx"    "ps"         
#> [51] "purrr"       "ragg"        "rappdirs"    "rex"         "rlang"      
#> [56] "rmarkdown"   "rprojroot"   "rstudioapi"  "sass"        "sessioninfo"
#> [61] "stringi"     "sys"         "systemfonts" "testthat"    "textshaping"
#> [66] "tibble"      "tinytex"     "usethis"     "utf8"        "vctrs"      
#> [71] "versionsort" "waldo"       "whisker"     "withr"       "woodendesc" 
#> [76] "xfun"        "xml2"        "yaml"        "zip"        
wood_local_packages("all")
#>   [1] "KernSmooth"   "MASS"         "Matrix"       "R6"           "_cache"      
#>   [6] "askpass"      "base"         "base64enc"    "boot"         "brio"        
#>  [11] "bslib"        "cachem"       "callr"        "class"        "cli"         
#>  [16] "clipr"        "cluster"      "codetools"    "compiler"     "covr"        
#>  [21] "crayon"       "credentials"  "curl"         "datasets"     "desc"        
#>  [26] "diffobj"      "digest"       "downlit"      "evaluate"     "fansi"       
#>  [31] "fastmap"      "fontawesome"  "foreign"      "fs"           "gert"        
#>  [36] "gh"           "gitcreds"     "glue"         "grDevices"    "graphics"    
#>  [41] "grid"         "highr"        "htmltools"    "httptest2"    "httr"        
#>  [46] "httr2"        "ini"          "jquerylib"    "jsonlite"     "knitr"       
#>  [51] "lattice"      "lazyeval"     "lifecycle"    "magrittr"     "memoise"     
#>  [56] "methods"      "mgcv"         "mime"         "nlme"         "nnet"        
#>  [61] "openssl"      "pak"          "parallel"     "pillar"       "pkgbuild"    
#>  [66] "pkgconfig"    "pkgdown"      "pkgload"      "praise"       "processx"    
#>  [71] "ps"           "purrr"        "ragg"         "rappdirs"     "rex"         
#>  [76] "rlang"        "rmarkdown"    "rpart"        "rprojroot"    "rstudioapi"  
#>  [81] "sass"         "sessioninfo"  "spatial"      "splines"      "stats"       
#>  [86] "stats4"       "stringi"      "survival"     "sys"          "systemfonts" 
#>  [91] "tcltk"        "testthat"     "textshaping"  "tibble"       "tinytex"     
#>  [96] "tools"        "translations" "usethis"      "utf8"         "utils"       
#> [101] "vctrs"        "versionsort"  "waldo"        "whisker"      "withr"       
#> [106] "woodendesc"   "xfun"         "xml2"         "yaml"         "zip"