Collect available packages from multiple repositories

This function queries the selected repositories for the available packages and collects the unique occurrences.

Usage

wood_packages(repos = "cran")

Arguments

repos

character()
A vector of repositories to query. The following values are available:

"cran", meaning CRAN;

"bioc@release", meaning Bioconductor, where release should be replaced with a valid Bioconductor release code or one of "release", "devel";

"github/user", meaning GitHub, where user should be replaced with a valid user or organization name (e.g. turtletopia);

"runiverse@universe", meaning R-universe, where universe should be replaced with a valid universe name (e.g. turtletopia);

"local#index", meaning a local library, where index should be replaced with an integer index of a library in base::.libPaths() or an all keyword;

"core", meaning base R packages;

a URL if none of the above is matched.

Value

A character vector of available packages.

See also

Examples

# \donttest{
# head() used due to the number of packages in there
head(wood_packages(c("local#all", "bioc@1.7", "cran", "core")))
#> [1] "A3"          "AATtools"    "ABACUS"      "ABC.RAP"     "ABCanalysis"
#> [6] "ABCoptim"   
wood_packages(c("http://www.omegahat.net/R", "runiverse@turtletopia"))
#>  [1] "Aspell"              "CGIwithR"            "CodeAnalysis"       
#>  [4] "CodeDepends"         "Combinations"        "FastCSVSample"      
#>  [7] "GeoIP"               "R2GoogleMaps"        "RAmazonDBREST"      
#> [10] "RAmazonS3"           "RAutoGenRunTime"     "RCIndex"            
#> [13] "RCSS"                "RCUDA"               "RCurl"              
#> [16] "RD3Device"           "RDCOMClient"         "REuPathDB"          
#> [19] "RExcelXML"           "RGCCTUFFI"           "RGCCTranslationUnit"
#> [22] "RGoogleDocs"         "RGoogleStorage"      "RGoogleTrends"      
#> [25] "RGraphicsDevice"     "RHTMLForms"          "RJSONIO"            
#> [28] "RKML"                "RKMLDevice"          "RNYTimes"           
#> [31] "ROOXML"              "ROpenOffice"         "RSPARQL"            
#> [34] "RSQLiteUDF"          "RSQLiteVirtualTable" "RTidyHTML"          
#> [37] "RUbigraph"           "RWordPress"          "RWordXML"           
#> [40] "Rcompression"        "Rffi"                "Rflickr"            
#> [43] "Rlibstemmer"         "Rnvvm"               "Rstem"              
#> [46] "SSOAP"               "SVGAnnotation"       "SWinTypeLibs"       
#> [49] "SpiderMonkey"        "Sxslt"               "WADL"               
#> [52] "XML"                 "XMLRPC"              "XMLSchema"          
#> [55] "Zillow"              "aurrera"             "deepdep"            
#> [58] "gglgbtq"             "ggpizza"             "openaccess"         
#> [61] "versionsort"         "woodendesc"         
# By default, only CRAN is queried:
head(wood_packages())
#> [1] "A3"          "AATtools"    "ABACUS"      "ABC.RAP"     "ABCanalysis"
#> [6] "ABCoptim"   
# }