Get dependencies of a package in one of R universes

This function queries the selected universe for dependencies of the selected package.

Usage

wood_runiverse_dependencies(package, universe = "ropensci")

Arguments

package

character(1)
Name of a package to query for.

universe

character(1)
Name of a universe within R-universe, e.g. "ropensci".

Value

A data frame with three columns, all in string format:

package (package name),

version (minimum version requirement or NA if none),

type (dependency type, e.g. "Imports").

See also

Examples

wood_runiverse_dependencies("targets")
#> <dependencies>
#>   Depends:     R (>= 3.5.0)
#>   Imports:     base64url (>= 1.4)
#>   Imports:     callr (>= 3.7.0)
#>   Imports:     cli (>= 2.0.2)
#>   Imports:     codetools (>= 0.2.16)
#>   Imports:     data.table (>= 1.12.8)
#>   Imports:     igraph (>= 2.0.0)
#>   Imports:     knitr (>= 1.34)
#>   Imports:     ps (>= 1.8.0)
#>   Imports:     R6 (>= 2.4.1)
#>   Imports:     rlang (>= 1.0.0)
#>   Imports:     secretbase (>= 0.5.0)
#>   Imports:     stats
#>   Imports:     tibble (>= 3.0.1)
#>   Imports:     tidyselect (>= 1.1.0)
#>   Imports:     tools
#>   Imports:     utils
#>   Imports:     vctrs (>= 0.2.4)
#>   Imports:     yaml (>= 2.2.1)
#>   Suggests:    autometric (>= 0.1.0)
#>   Suggests:    bslib
#>   Suggests:    clustermq (>= 0.9.2)
#>   Suggests:    crew (>= 0.9.0)
#>   Suggests:    curl (>= 4.3)
#>   Suggests:    DT (>= 0.14)
#>   Suggests:    dplyr (>= 1.0.0)
#>   Suggests:    fst (>= 0.9.2)
#>   Suggests:    future (>= 1.19.1)
#>   Suggests:    future.batchtools (>= 0.9.0)
#>   Suggests:    future.callr (>= 0.6.0)
#>   Suggests:    gargle (>= 1.2.0)
#>   Suggests:    googleCloudStorageR (>= 0.7.0)
#>   Suggests:    gt (>= 0.2.2)
#>   Suggests:    keras (>= 2.2.5.0)
#>   Suggests:    markdown (>= 1.1)
#>   Suggests:    nanonext (>= 0.12.0)
#>   Suggests:    rmarkdown (>= 2.4)
#>   Suggests:    parallelly (>= 1.35.0)
#>   Suggests:    paws.common (>= 0.6.4)
#>   Suggests:    paws.storage (>= 0.4.0)
#>   Suggests:    pkgload (>= 1.1.0)
#>   Suggests:    processx (>= 3.4.3)
#>   Suggests:    qs2
#>   Suggests:    reprex (>= 2.0.0)
#>   Suggests:    rstudioapi (>= 0.11)
#>   Suggests:    R.utils (>= 2.6.0)
#>   Suggests:    shiny (>= 1.5.0)
#>   Suggests:    shinybusy (>= 0.2.2)
#>   Suggests:    shinyWidgets (>= 0.5.4)
#>   Suggests:    tarchetypes
#>   Suggests:    testthat (>= 3.0.0)
#>   Suggests:    torch (>= 0.1.0)
#>   Suggests:    usethis (>= 1.6.3)
#>   Suggests:    visNetwork (>= 2.1.2)
wood_runiverse_dependencies("ggplot2", universe = "tidyverse")
#> <dependencies>
#>   Depends:     R (>= 4.0)
#>   Imports:     cli
#>   Imports:     grDevices
#>   Imports:     grid
#>   Imports:     gtable (>= 0.1.1)
#>   Imports:     isoband
#>   Imports:     lifecycle
#>   Imports:     rlang (>= 1.1.0)
#>   Imports:     scales (>= 1.3.0)
#>   Imports:     stats
#>   Imports:     vctrs (>= 0.6.0)
#>   Imports:     withr (>= 2.5.0)
#>   Suggests:    broom
#>   Suggests:    covr
#>   Suggests:    dplyr
#>   Suggests:    ggplot2movies
#>   Suggests:    hexbin
#>   Suggests:    Hmisc
#>   Suggests:    knitr
#>   Suggests:    mapproj
#>   Suggests:    maps
#>   Suggests:    MASS
#>   Suggests:    mgcv
#>   Suggests:    multcomp
#>   Suggests:    munsell
#>   Suggests:    nlme
#>   Suggests:    profvis
#>   Suggests:    quantreg
#>   Suggests:    ragg (>= 1.2.6)
#>   Suggests:    RColorBrewer
#>   Suggests:    rmarkdown
#>   Suggests:    rpart
#>   Suggests:    sf (>= 0.7-3)
#>   Suggests:    svglite (>= 2.1.2)
#>   Suggests:    testthat (>= 3.1.5)
#>   Suggests:    tibble
#>   Suggests:    vdiffr (>= 1.0.6)
#>   Suggests:    xml2
#>   Enhances:    sp