List available packages in one of R universes

This function queries the selected universe for a list of available packages. They are returned as a vector of strings, each element being a package name.

Usage

wood_runiverse_packages(universe = "ropensci")

Arguments

universe

character(1)
Name of a universe within R-universe, e.g. "ropensci".

Value

A character vector of available packages.

See also

Examples

# \donttest{
wood_runiverse_packages()
#>   [1] "BaseSet"            "CRediTas"           "CoordinateCleaner" 
#>   [4] "DataPackageR"       "DataSpaceR"         "DoOR.data"         
#>   [7] "DoOR.functions"     "EDIutils"           "EML"               
#>  [10] "EndoMineR"          "FedData"            "GSODR"             
#>  [13] "MODIStsp"           "MtreeRing"          "NLMR"              
#>  [16] "PostcodesioR"       "RNeXML"             "RSelenium"         
#>  [19] "ReLTER"             "RefManageR"         "Rpolyhedra"        
#>  [22] "SBC"                "SymbiotaR2"         "UCSCXenaTools"     
#>  [25] "USAboundaries"      "USAboundariesData"  "aRxiv"             
#>  [28] "allcontributors"    "allodb"             "antanym"           
#>  [31] "antiword"           "aorsf"              "arkdb"             
#>  [34] "assertr"            "auk"                "autotest"          
#>  [37] "av"                 "awardFindR"         "baRcodeR"          
#>  [40] "babeldown"          "babette"            "beastier"          
#>  [43] "beautier"           "bib2df"             "bibtex"            
#>  [46] "bikedata"           "binman"             "biomartr"          
#>  [49] "bold"               "bowerbird"          "brranching"        
#>  [52] "c14bazAAR"          "cRegulome"          "camsRad"           
#>  [55] "canaper"            "cde"                "censo2017"         
#>  [58] "cffr"               "charlatan"          "chirps"            
#>  [61] "chlorpromazineR"    "chromer"            "circle"            
#>  [64] "citecorp"           "citesdb"            "ckanr"             
#>  [67] "cld2"               "cld3"               "clifro"            
#>  [70] "cmdstanr"           "codemetar"          "coder"             
#>  [73] "colocr"             "commonmark"         "comtradr"          
#>  [76] "concstats"          "credentials"        "crew"              
#>  [79] "crul"               "cyphr"              "daiquiri"          
#>  [82] "dataaimsr"          "datapack"           "dataspice"         
#>  [85] "datefixR"           "dbhydroR"           "dbparser"          
#>  [88] "deposits"           "dittodb"            "drake"             
#>  [91] "dynamite"           "eia"                "elastic"           
#>  [94] "emld"               "emo"                "epair"             
#>  [97] "epubr"              "essurvey"           "europepmc"         
#> [100] "excluder"           "exoplanets"         "ezknitr"           
#> [103] "fingertipsR"        "frictionless"       "gbifdb"            
#> [106] "genbankr"           "gendercoder"        "geojson"           
#> [109] "geojsonio"          "geonames"           "gert"              
#> [112] "getCRUCLdata"       "ggmap"              "ghql"              
#> [115] "gistr"              "git2r"              "git2rdata"         
#> [118] "gitcellar"          "gitignore"          "gittargets"        
#> [121] "goodpractice"       "googleLanguageR"    "grainchanger"      
#> [124] "graphql"            "gutenbergr"         "handlr"            
#> [127] "hddtools"           "helminthR"          "historydata"       
#> [130] "hoardr"             "hunspell"           "hydroscoper"       
#> [133] "icepalace"          "iheatmapr"          "ijtiff"            
#> [136] "internetarchive"    "jagstargets"        "jenkins"           
#> [139] "jqr"                "jsonld"             "jsonvalidate"      
#> [142] "jstor"              "katex"              "landscapetools"    
#> [145] "leafem"             "lightr"             "lingtypology"      
#> [148] "lubridate"          "magick"             "mapscanner"        
#> [151] "mapview"            "mauricer"           "mcbette"           
#> [154] "mctq"               "medrxivr"           "melt"              
#> [157] "mirai"              "mregions"           "nasapower"         
#> [160] "nbaR"               "neotoma"            "nlrx"              
#> [163] "nodbi"              "nomisr"             "npi"               
#> [166] "oai"                "occCite"            "onekp"             
#> [169] "openalexR"          "opencage"           "opencv"            
#> [172] "opentripplanner"    "osfr"               "osmdata"           
#> [175] "osmextract"         "osmplotr"           "outcomerate"       
#> [178] "paleobioDB"         "pangaear"           "parzer"            
#> [181] "patentsview"        "pathviewr"          "pdftools"          
#> [184] "phonfieldwork"      "photosearcher"      "phruta"            
#> [187] "phylocomr"          "phylogram"          "phylotaR"          
#> [190] "piggyback"          "pixelclasser"       "pkgcheck"          
#> [193] "pkgreviewr"         "pkgstats"           "plater"            
#> [196] "popler"             "postdoc"            "predictNMB"        
#> [199] "prism"              "prismjs"            "qcoder"            
#> [202] "qpdf"               "qualR"              "qualtRics"         
#> [205] "quartificate"       "r2readthedocs"      "rWBclimate"        
#> [208] "ramlegacy"          "rb3"                "rcites"            
#> [211] "rcrossref"          "rdatacite"          "rdataretriever"    
#> [214] "rdefra"             "rdflib"             "rdhs"              
#> [217] "rdryad"             "readODS"            "rebird"            
#> [220] "redland"            "refsplitr"          "rentrez"           
#> [223] "rerddap"            "restez"             "rfema"             
#> [226] "rfishbase"          "rfisheries"         "rgbif"             
#> [229] "rglobi"             "rgnparser"          "rgpdd"             
#> [232] "riem"               "rinat"              "ritis"             
#> [235] "rmangal"            "rnassqs"            "rnaturalearth"     
#> [238] "rnaturalearthdata"  "rnaturalearthhires" "rnoaa"             
#> [241] "roadoi"             "roblog"             "robotstxt"         
#> [244] "roreviewapi"        "rotemplate"         "rotl"              
#> [247] "rperseus"           "rppo"               "rredlist"          
#> [250] "rrlite"             "rrricanes"          "rrricanesdata"     
#> [253] "rsat"               "rsnps"              "rsvg"              
#> [256] "rtika"              "rtweet"             "ruODK"             
#> [259] "rusda"              "rzmq"               "skimr"             
#> [262] "skynet"             "slopes"             "smapr"             
#> [265] "sodium"             "sofa"               "spatsoc"           
#> [268] "spelling"           "spiro"              "spocc"             
#> [271] "srr"                "ssh"                "stantargets"       
#> [274] "stats19"            "stplanr"            "suppdata"          
#> [277] "tabulizer"          "tabulizerjars"      "tacmagic"          
#> [280] "tarchetypes"        "targets"            "taxa"              
#> [283] "taxadb"             "taxize"             "taxizedb"          
#> [286] "taxlist"            "terrainr"           "tesseract"         
#> [289] "textreuse"          "tic"                "tidyhydat"         
#> [292] "tidync"             "tidypmc"            "tidyqpcr"          
#> [295] "tidytags"           "tif"                "tiler"             
#> [298] "tinkr"              "tokenizers"         "tracerer"          
#> [301] "tradestatistics"    "traits"             "treebase"          
#> [304] "treedata.table"     "treeio"             "treeman"           
#> [307] "treestartr"         "tsbox"              "unifir"            
#> [310] "universe"           "unrtf"              "vcr"               
#> [313] "virtuoso"           "visdat"             "wateRinfo"         
#> [316] "waywiser"           "wdman"              "weathercan"        
#> [319] "webchem"            "webmockr"           "wikitaxa"          
#> [322] "workloopR"          "worrms"             "writexl"           
#> [325] "xslt"               "yfR"               
wood_runiverse_packages("tidyverse")
#>  [1] "blob"          "dbplyr"        "dplyr"         "dtplyr"       
#>  [5] "forcats"       "ggplot2"       "glue"          "googledrive"  
#>  [9] "googlesheets4" "haven"         "hms"           "lubridate"    
#> [13] "magrittr"      "modelr"        "multidplyr"    "nycflights13" 
#> [17] "purrr"         "readr"         "readxl"        "reprex"       
#> [21] "rvest"         "stringr"       "tibble"        "tidyr"        
#> [25] "tidyverse"     "vroom"        
# }