Get dependencies of a package in any repository

This function queries any online repository for dependencies of the selected package.

Usage

wood_url_dependencies(package, repository)

Arguments

package

character(1)
Name of a package to query for.

repository

character(1)
URL to repository, e.g. "http://www.omegahat.net/R".

Value

A data frame with three columns, all in string format:

package (package name),

version (minimum version requirement or NA if none),

type (dependency type, e.g. "Imports").

See also

Examples

wood_url_dependencies("tidystringdist", repository = "https://colinfay.me")
#> <dependencies>
#>   Imports:     stringdist
#>   Imports:     rlang
#>   Imports:     magrittr
#>   Imports:     dplyr
#>   Imports:     purrr
#>   Imports:     attempt
#>   Suggests:    testthat
#>   Suggests:    knitr
#>   Suggests:    rmarkdown
#>   Suggests:    covr