Find dependencies from multiple repositories

This function queries the selected repositories for dependencies of the selected packages and returns the first working occurence.

Usage

wood_dependencies(packages, repos = "cran")

Arguments

packages

character()
Names of the packages to query for.

repos

character()
A vector of repositories to query. The following values are available:

"cran", meaning CRAN;

"bioc@release", meaning Bioconductor, where release should be replaced with a valid Bioconductor release code or one of "release", "devel";

"core", meaning base R packages;

"runiverse@universe", meaning R-universe, where universe should be replaced with a valid universe name (e.g. turtletopia);

"local#index", meaning a local library, where index should be replaced with an integer index of a library in base::.libPaths() or an all keyword;

a URL if none of the above is matched.

Value

A list named after queried packages, each element being a data frame with three columns, all in string format:

package (package name),

version (minimum version requirement or NA if none),

type (dependency type, e.g. "Imports").

See also

Examples

# \donttest{
wood_dependencies("stats", "core")
#> <woodendesc dependency list [1]>
#>  <stats dependencies>
#>   Imports:     utils
#>   Imports:     grDevices
#>   Imports:     graphics
#>   Suggests:    MASS
#>   Suggests:    Matrix
#>   Suggests:    SuppDists
#>   Suggests:    methods
#>   Suggests:    stats4
#> 
# Multiple packages are also possible:
wood_dependencies(
  c("ggplot2", "Biostrings", "woodendesc"),
  repos = c("runiverse@turtletopia", "cran", "bioc@1.5")
)
#> <woodendesc dependency list [3]>
#>  <ggplot2 dependencies>
#>   Depends:     R (>= 3.3)
#>   Imports:     cli
#>   Imports:     glue
#>   Imports:     grDevices
#>   Imports:     grid
#>   Imports:     gtable (>= 0.1.1)
#>   Imports:     isoband
#>   Imports:     lifecycle
#>   Imports:     MASS
#>   Imports:     mgcv
#>   Imports:     rlang (>= 1.0.0)
#>   Imports:     scales (>= 1.2.0)
#>   Imports:     stats
#>   Imports:     tibble
#>   Imports:     vctrs (>= 0.5.0)
#>   Imports:     withr (>= 2.5.0)
#>   Suggests:    covr
#>   Suggests:    dplyr
#>   Suggests:    ggplot2movies
#>   Suggests:    hexbin
#>   Suggests:    Hmisc
#>   Suggests:    knitr
#>   Suggests:    lattice
#>   Suggests:    mapproj
#>   Suggests:    maps
#>   Suggests:    maptools
#>   Suggests:    multcomp
#>   Suggests:    munsell
#>   Suggests:    nlme
#>   Suggests:    profvis
#>   Suggests:    quantreg
#>   Suggests:    ragg
#>   Suggests:    RColorBrewer
#>   Suggests:    rgeos
#>   Suggests:    rmarkdown
#>   Suggests:    rpart
#>   Suggests:    sf (>= 0.7-3)
#>   Suggests:    svglite (>= 1.2.0.9001)
#>   Suggests:    testthat (>= 3.1.2)
#>   Suggests:    vdiffr (>= 1.0.0)
#>   Suggests:    xml2
#>   Enhances:    sp
#> 
#>  <Biostrings dependencies>
#>   Depends:     R (>= 1.8.0)
#> 
#>  <woodendesc dependencies>
#>   Depends:     R (>= 3.5.0)
#>   Imports:     digest
#>   Imports:     httr
#>   Imports:     versionsort (>= 1.1.0)
#>   Suggests:    covr
#>   Suggests:    httptest
#>   Suggests:    testthat (>= 3.0.0)
#>   Suggests:    usethis
#>   Suggests:    vcr
#>   Suggests:    withr
#>   Suggests:    xml2
#> 
# By default, only CRAN is queried:
wood_dependencies("versionsort")
#> <woodendesc dependency list [1]>
#>  <versionsort dependencies>
#>   Suggests:    covr
#>   Suggests:    spelling
#>   Suggests:    testthat (>= 3.0.0)
#> 
# }