List locally available packages

This function searches for installed packages inside specified library paths.

Usage

wood_local_packages(paths = .libPaths()[1])

Arguments

paths

character()
Paths to local libraries, by default the first element of .libPaths(). If equal to "all", uses all elements of .libPaths().

Value

A character vector of available packages.

See also

Examples

wood_local_packages()
#>  [1] "R6"          "Rcpp"        "_cache"      "askpass"     "base64enc"  
#>  [6] "brio"        "bslib"       "cachem"      "callr"       "cli"        
#> [11] "clipr"       "covr"        "cpp11"       "crayon"      "credentials"
#> [16] "crul"        "curl"        "desc"        "diffobj"     "digest"     
#> [21] "downlit"     "ellipsis"    "evaluate"    "fansi"       "fastmap"    
#> [26] "fauxpas"     "fs"          "gert"        "gh"          "gitcreds"   
#> [31] "glue"        "highr"       "htmltools"   "httpcode"    "httptest"   
#> [36] "httr"        "httr2"       "ini"         "jquerylib"   "jsonlite"   
#> [41] "knitr"       "lazyeval"    "lifecycle"   "magrittr"    "memoise"    
#> [46] "mime"        "openssl"     "pillar"      "pkgbuild"    "pkgconfig"  
#> [51] "pkgdown"     "pkgload"     "praise"      "prettyunits" "processx"   
#> [56] "ps"          "purrr"       "ragg"        "rappdirs"    "rcmdcheck"  
#> [61] "rematch2"    "rex"         "rlang"       "rmarkdown"   "rprojroot"  
#> [66] "rstudioapi"  "sass"        "sessioninfo" "stringi"     "stringr"    
#> [71] "sys"         "systemfonts" "testthat"    "textshaping" "tibble"     
#> [76] "tinytex"     "triebeard"   "urltools"    "usethis"     "utf8"       
#> [81] "vcr"         "vctrs"       "versionsort" "waldo"       "webmockr"   
#> [86] "whisker"     "withr"       "woodendesc"  "xfun"        "xml2"       
#> [91] "xopen"       "yaml"        "zip"        
wood_local_packages("all")
#>   [1] "KernSmooth"   "MASS"         "Matrix"       "R6"           "Rcpp"        
#>   [6] "_cache"       "askpass"      "base"         "base64enc"    "boot"        
#>  [11] "brio"         "bslib"        "cachem"       "callr"        "class"       
#>  [16] "cli"          "clipr"        "cluster"      "codetools"    "compiler"    
#>  [21] "covr"         "cpp11"        "crayon"       "credentials"  "crul"        
#>  [26] "curl"         "datasets"     "desc"         "diffobj"      "digest"      
#>  [31] "downlit"      "ellipsis"     "evaluate"     "fansi"        "fastmap"     
#>  [36] "fauxpas"      "foreign"      "fs"           "gert"         "gh"          
#>  [41] "gitcreds"     "glue"         "grDevices"    "graphics"     "grid"        
#>  [46] "highr"        "htmltools"    "httpcode"     "httptest"     "httr"        
#>  [51] "httr2"        "ini"          "jquerylib"    "jsonlite"     "knitr"       
#>  [56] "lattice"      "lazyeval"     "lifecycle"    "magrittr"     "memoise"     
#>  [61] "methods"      "mgcv"         "mime"         "nlme"         "nnet"        
#>  [66] "openssl"      "pak"          "parallel"     "pillar"       "pkgbuild"    
#>  [71] "pkgconfig"    "pkgdown"      "pkgload"      "praise"       "prettyunits" 
#>  [76] "processx"     "ps"           "purrr"        "ragg"         "rappdirs"    
#>  [81] "rcmdcheck"    "rematch2"     "rex"          "rlang"        "rmarkdown"   
#>  [86] "rpart"        "rprojroot"    "rstudioapi"   "sass"         "sessioninfo" 
#>  [91] "spatial"      "splines"      "stats"        "stats4"       "stringi"     
#>  [96] "stringr"      "survival"     "sys"          "systemfonts"  "tcltk"       
#> [101] "testthat"     "textshaping"  "tibble"       "tinytex"      "tools"       
#> [106] "translations" "triebeard"    "urltools"     "usethis"      "utf8"        
#> [111] "utils"        "vcr"          "vctrs"        "versionsort"  "waldo"       
#> [116] "webmockr"     "whisker"      "withr"        "woodendesc"   "xfun"        
#> [121] "xml2"         "xopen"        "yaml"         "zip"