List available packages in one of R universes

This function queries the selected universe for a list of available packages. They are returned as a vector of strings, each element being a package name.

Usage

wood_runiverse_packages(universe = "ropensci")

Arguments

universe

character(1)
Name of a universe within R-universe, e.g. "ropensci".

Value

A character vector of available packages.

See also

Examples

# \donttest{
wood_runiverse_packages()
#>   [1] "BaseSet"            "CoordinateCleaner"  "DataPackageR"      
#>   [4] "DataSpaceR"         "DoOR.data"          "DoOR.functions"    
#>   [7] "EDIutils"           "EML"                "EndoMineR"         
#>  [10] "FedData"            "GSODR"              "MODIStsp"          
#>  [13] "MtreeRing"          "NLMR"               "PostcodesioR"      
#>  [16] "RNeXML"             "RSelenium"          "ReLTER"            
#>  [19] "RefManageR"         "Rpolyhedra"         "SBC"               
#>  [22] "SymbiotaR2"         "UCSCXenaTools"      "USAboundaries"     
#>  [25] "USAboundariesData"  "aRxiv"              "allcontributors"   
#>  [28] "allodb"             "antanym"            "antiword"          
#>  [31] "aorsf"              "arkdb"              "assertr"           
#>  [34] "auk"                "autotest"           "av"                
#>  [37] "awardFindR"         "baRcodeR"           "babeldown"         
#>  [40] "babette"            "beastier"           "beautier"          
#>  [43] "bib2df"             "bibtex"             "bikedata"          
#>  [46] "binman"             "biomartr"           "bold"              
#>  [49] "bowerbird"          "brranching"         "c14bazAAR"         
#>  [52] "cRegulome"          "camsRad"            "canaper"           
#>  [55] "cde"                "censo2017"          "cffr"              
#>  [58] "charlatan"          "chirps"             "chlorpromazineR"   
#>  [61] "chromer"            "circle"             "citecorp"          
#>  [64] "citesdb"            "ckanr"              "cld2"              
#>  [67] "cld3"               "clifro"             "cmdstanr"          
#>  [70] "codemetar"          "coder"              "colocr"            
#>  [73] "commonmark"         "comtradr"           "credentials"       
#>  [76] "crul"               "cyphr"              "daiquiri"          
#>  [79] "dataaimsr"          "datapack"           "dataspice"         
#>  [82] "datefixR"           "dbhydroR"           "dbparser"          
#>  [85] "deposits"           "dittodb"            "drake"             
#>  [88] "dynamite"           "eia"                "elastic"           
#>  [91] "emld"               "emo"                "epair"             
#>  [94] "epubr"              "essurvey"           "europepmc"         
#>  [97] "excluder"           "exoplanets"         "ezknitr"           
#> [100] "fingertipsR"        "frictionless"       "gbifdb"            
#> [103] "genbankr"           "gendercoder"        "geojsonio"         
#> [106] "geonames"           "gert"               "getCRUCLdata"      
#> [109] "ggmap"              "ghql"               "gistr"             
#> [112] "git2r"              "git2rdata"          "gitcellar"         
#> [115] "gitignore"          "gittargets"         "goodpractice"      
#> [118] "googleLanguageR"    "grainchanger"       "graphql"           
#> [121] "gutenbergr"         "handlr"             "hddtools"          
#> [124] "helminthR"          "historydata"        "hoardr"            
#> [127] "hunspell"           "hydroscoper"        "icepalace"         
#> [130] "iheatmapr"          "ijtiff"             "internetarchive"   
#> [133] "jagstargets"        "jenkins"            "jqr"               
#> [136] "jsonld"             "jsonvalidate"       "jstor"             
#> [139] "katex"              "landscapetools"     "leafem"            
#> [142] "lightr"             "lingtypology"       "lubridate"         
#> [145] "magick"             "mapscanner"         "mapview"           
#> [148] "mauricer"           "mcbette"            "mctq"              
#> [151] "medrxivr"           "melt"               "mregions"          
#> [154] "nasapower"          "nbaR"               "neotoma"           
#> [157] "nlrx"               "nodbi"              "nomisr"            
#> [160] "npi"                "oai"                "occCite"           
#> [163] "onekp"              "openalexR"          "opencage"          
#> [166] "opencv"             "opentripplanner"    "osfr"              
#> [169] "osmdata"            "osmextract"         "osmplotr"          
#> [172] "outcomerate"        "paleobioDB"         "pangaear"          
#> [175] "parzer"             "patentsview"        "pathviewr"         
#> [178] "pdftools"           "phonfieldwork"      "photosearcher"     
#> [181] "phruta"             "phylocomr"          "phylogram"         
#> [184] "phylotaR"           "piggyback"          "pixelclasser"      
#> [187] "pkgcheck"           "pkgreviewr"         "pkgstats"          
#> [190] "plater"             "popler"             "postdoc"           
#> [193] "prism"              "prismjs"            "qcoder"            
#> [196] "qpdf"               "qualR"              "qualtRics"         
#> [199] "r2readthedocs"      "rWBclimate"         "ramlegacy"         
#> [202] "rb3"                "rcites"             "rcrossref"         
#> [205] "rdatacite"          "rdataretriever"     "rdefra"            
#> [208] "rdflib"             "rdhs"               "rdryad"            
#> [211] "readODS"            "rebird"             "redland"           
#> [214] "refsplitr"          "rentrez"            "rerddap"           
#> [217] "restez"             "rfema"              "rfishbase"         
#> [220] "rfisheries"         "rgbif"              "rglobi"            
#> [223] "rgnparser"          "rgpdd"              "riem"              
#> [226] "rinat"              "ritis"              "rmangal"           
#> [229] "rnassqs"            "rnaturalearth"      "rnaturalearthdata" 
#> [232] "rnaturalearthhires" "rnoaa"              "roadoi"            
#> [235] "roblog"             "robotstxt"          "roreviewapi"       
#> [238] "rotemplate"         "rotl"               "rperseus"          
#> [241] "rppo"               "rredlist"           "rrlite"            
#> [244] "rrricanes"          "rrricanesdata"      "rsat"              
#> [247] "rsnps"              "rsvg"               "rtika"             
#> [250] "rtweet"             "ruODK"              "rusda"             
#> [253] "rzmq"               "skimr"              "skynet"            
#> [256] "slopes"             "smapr"              "sodium"            
#> [259] "spatsoc"            "spelling"           "spiro"             
#> [262] "spocc"              "srr"                "ssh"               
#> [265] "stantargets"        "stats19"            "stplanr"           
#> [268] "suppdata"           "tabulizer"          "tabulizerjars"     
#> [271] "tacmagic"           "tarchetypes"        "targets"           
#> [274] "taxa"               "taxadb"             "taxize"            
#> [277] "taxizedb"           "taxlist"            "terrainr"          
#> [280] "tesseract"          "textreuse"          "tic"               
#> [283] "tidyhydat"          "tidync"             "tidypmc"           
#> [286] "tidyqpcr"           "tidytags"           "tif"               
#> [289] "tiler"              "tinkr"              "tokenizers"        
#> [292] "tracerer"           "tradestatistics"    "traits"            
#> [295] "treebase"           "treedata.table"     "treeio"            
#> [298] "treeman"            "treestartr"         "unifir"            
#> [301] "universe"           "unrtf"              "vcr"               
#> [304] "virtuoso"           "visdat"             "wateRinfo"         
#> [307] "waywiser"           "wdman"              "weathercan"        
#> [310] "webchem"            "webmockr"           "wellknown"         
#> [313] "wikitaxa"           "workloopR"          "worrms"            
#> [316] "writexl"            "xslt"               "yfR"               
wood_runiverse_packages("tidyverse")
#>  [1] "blob"          "dbplyr"        "dplyr"         "dtplyr"       
#>  [5] "forcats"       "gargle"        "ggplot2"       "glue"         
#>  [9] "googledrive"   "googlesheets4" "haven"         "hms"          
#> [13] "lubridate"     "magrittr"      "modelr"        "multidplyr"   
#> [17] "nycflights13"  "purrr"         "readr"         "readxl"       
#> [21] "reprex"        "rlang"         "rvest"         "stringr"      
#> [25] "tibble"        "tidyr"         "tidyverse"     "vctrs"        
#> [29] "vroom"        
# }