List available packages in one of R universes

This function queries the selected universe for a list of available packages. They are returned as a vector of strings, each element being a package name.

Usage

wood_runiverse_packages(universe = "ropensci")

Arguments

universe

character(1)
Name of a universe within R-universe, e.g. "ropensci".

Value

A character vector of available packages.

See also

Examples

wood_runiverse_packages()
#>   [1] "rredlist"           "targets"            "av"                
#>   [4] "emodnet.wfs"        "rix"                "ijtiff"            
#>   [7] "openalexR"          "cffr"               "ows4R"             
#>  [10] "beautier"           "SBC"                "traits"            
#>  [13] "opencage"           "eDNAjoint"          "skimr"             
#>  [16] "git2rdata"          "baRulho"            "mregions2"         
#>  [19] "pkgmatch"           "rsi"                "GSODR"             
#>  [22] "chirps"             "osmextract"         "riem"              
#>  [25] "gigs"               "hoardr"             "taxize"            
#>  [28] "git2r"              "cmdstanr"           "cyphr"             
#>  [31] "stats19"            "readODS"            "leafem"            
#>  [34] "dendroNetwork"      "gutenbergr"         "circle"            
#>  [37] "tarchetypes"        "rerddap"            "prismjs"           
#>  [40] "bowerbird"          "mctq"               "auk"               
#>  [43] "tabulapdf"          "rfishbase"          "patentsview"       
#>  [46] "allcontributors"    "webchem"            "opentripplanner"   
#>  [49] "mapview"            "babeldown"          "qualR"             
#>  [52] "aeolus"             "roblog"             "babelquarto"       
#>  [55] "nasapower"          "getCRUCLdata"       "rzmq"              
#>  [58] "xslt"               "jqr"                "sodium"            
#>  [61] "webmockr"           "bold"               "rnaturalearth"     
#>  [64] "biomartr"           "FedData"            "excluder"          
#>  [67] "spatsoc"            "datefixR"           "citecorp"          
#>  [70] "rdhs"               "agroclimatico"      "drake"             
#>  [73] "jagstargets"        "stantargets"        "internetarchive"   
#>  [76] "lightr"             "weatherOz"          "rebird"            
#>  [79] "pkgcheck"           "piggyback"          "photosearcher"     
#>  [82] "chopin"             "roreviewapi"        "RefManageR"        
#>  [85] "goodpractice"       "osmapiR"            "dynamite"          
#>  [88] "robotstxt"          "comtradr"           "weathercan"        
#>  [91] "tinkr"              "osmdata"            "nodbi"             
#>  [94] "ruODK"              "medrxivr"           "pkgstats"          
#>  [97] "rangr"              "rrlite"             "GLMMcosinor"       
#> [100] "beastier"           "gert"               "gitignore"         
#> [103] "Rpolyhedra"         "rOPTRAM"            "treeio"            
#> [106] "UCSCXenaTools"      "lingtypology"       "occCite"           
#> [109] "srr"                "unrtf"              "plater"            
#> [112] "autotest"           "mapscanner"         "ohun"              
#> [115] "universe"           "osmplotr"           "tsbox"             
#> [118] "rvertnet"           "charlatan"          "rgbif"             
#> [121] "tidync"             "cde"                "commonmark"        
#> [124] "wikitaxa"           "naijR"              "slopes"            
#> [127] "tidyhydat"          "nuts"               "credentials"       
#> [130] "fluidsynth"         "hunspell"           "qpdf"              
#> [133] "writexl"            "cld2"               "cld3"              
#> [136] "graphql"            "opencv"             "spelling"          
#> [139] "antiword"           "tesseract"          "jsonld"            
#> [142] "postdoc"            "katex"              "spiro"             
#> [145] "frictionless"       "roadoi"             "pdftools"          
#> [148] "rsvg"               "magick"             "ssh"               
#> [151] "rnaturalearthhires" "censo2017"          "DataPackageR"      
#> [154] "taxlist"            "skynet"             "epubr"             
#> [157] "tiler"              "qualtRics"          "jenkins"           
#> [160] "mauricer"           "deposits"           "nlrx"              
#> [163] "daiquiri"           "hddtools"           "EndoMineR"         
#> [166] "NLMR"               "nomisr"             "rotemplate"        
#> [169] "rnassqs"            "tradestatistics"    "stplanr"           
#> [172] "historydata"        "refsplitr"          "rb3"               
#> [175] "iheatmapr"          "bib2df"             "eph"               
#> [178] "rdflib"             "textreuse"          "rsgeo"             
#> [181] "mcbette"            "phonfieldwork"      "vcr"               
#> [184] "karel"              "rdataretriever"     "crul"              
#> [187] "predictNMB"         "eia"                "dwctaxon"          
#> [190] "phylotaR"           "hydroscoper"        "MODIStsp"          
#> [193] "visdat"             "dbparser"           "gittargets"        
#> [196] "googleLanguageR"    "workloopR"          "tic"               
#> [199] "jstor"              "MtreeRing"          "waywiser"          
#> [202] "rfema"              "aorsf"              "terrainr"          
#> [205] "phruta"             "codemetar"          "osfr"              
#> [208] "yfR"                "babette"            "r2readthedocs"     
#> [211] "taxizedb"           "c14bazAAR"          "gitcellar"         
#> [214] "sofa"               "fellingdater"       "melt"              
#> [217] "RNeXML"             "tidyqpcr"           "virtuoso"          
#> [220] "restez"             "rsat"               "taxadb"            
#> [223] "assertr"            "dittodb"            "chromer"           
#> [226] "tokenizers"         "jsonvalidate"       "awardFindR"        
#> [229] "quadkeyr"           "concstats"          "spocc"             
#> [232] "CoordinateCleaner"  "aRxiv"              "paleobioDB"        
#> [235] "ggmap"              "redland"            "gendercoder"       
#> [238] "fingertipsR"        "units"              "taxa"              
#> [241] "DoOR.functions"     "treebase"           "rnaturalearthdata" 
#> [244] "bikedata"           "unifir"             "baRcodeR"          
#> [247] "natserv"            "arkdb"              "worrms"            
#> [250] "rmangal"            "ReLTER"             "BaseSet"           
#> [253] "rgnparser"          "birdsize"           "prism"             
#> [256] "tif"                "bibtex"             "europepmc"         
#> [259] "fastMatMR"          "geojsonio"          "gbifdb"            
#> [262] "suppdata"           "wdman"              "EDIutils"          
#> [265] "rrricanes"          "wmm"                "rglobi"            
#> [268] "tracerer"           "pkgreviewr"         "ezknitr"           
#> [271] "geojson"            "binman"             "RSelenium"         
#> [274] "rsnps"              "rppo"               "rotl"              
#> [277] "rperseus"           "rnoaa"              "grainchanger"      
#> [280] "dataaimsr"          "elastic"            "canaper"           
#> [283] "CRediTas"           "rtika"              "phylocomr"         
#> [286] "PostcodesioR"       "onekp"              "quartificate"      
#> [289] "ckanr"              "rcrossref"          "coder"             
#> [292] "clifro"             "pathviewr"          "parzer"            
#> [295] "smapr"              "rcites"             "rdryad"            
#> [298] "ghql"               "rdatacite"          "helminthR"         
#> [301] "brranching"         "neotoma"            "pangaear"          
#> [304] "npi"                "oai"                "gistr"             
#> [307] "landscapetools"     "ritis"              "exoplanets"        
#> [310] "antanym"            "epair"              "wateRinfo"         
#> [313] "DataSpaceR"         "rinat"              "datapack"          
#> [316] "EML"                "icepalace"          "handlr"            
#> [319] "DoOR.data"          "SymbiotaR2"         "essurvey"          
#> [322] "qcoder"             "allodb"             "USAboundaries"     
#> [325] "USAboundariesData"  "treedata.table"     "dataspice"         
#> [328] "chlorpromazineR"    "pixelclasser"       "rusda"             
#> [331] "outcomerate"        "rentrez"            "emld"              
#> [334] "rgpdd"              "cRegulome"          "tacmagic"          
#> [337] "colocr"             "emo"                "rfisheries"        
#> [340] "phylogram"          "tidypmc"            "rrricanesdata"     
#> [343] "ramlegacy"          "camsRad"            "treestartr"        
#> [346] "popler"             "geonames"          
wood_runiverse_packages("tidyverse")
#>  [1] "purrr"         "ggplot2"       "duckplyr"      "ellmer"       
#>  [5] "dtplyr"        "tibble"        "lubridate"     "hms"          
#>  [9] "blob"          "dbplyr"        "dplyr"         "rvest"        
#> [13] "tidyr"         "tidyverse"     "haven"         "glue"         
#> [17] "reprex"        "readxl"        "vroom"         "stringr"      
#> [21] "googledrive"   "readr"         "multidplyr"    "googlesheets4"
#> [25] "nycflights13"  "forcats"       "modelr"        "magrittr"